almost work domain ssl update

+++ frontend/app.py
@@ -20,6 +20,8 @@ import subprocess
 from pydantic import BaseModel
 from pathlib import Path
 
+import asyncio
+
 from clickhouse_connect import get_client
 
 app = FastAPI()
@@ -192,6 +194,11 @@ def show_logs():
     return TRACK_LOG[-50:]  # последние 50 событий
 
 
+@app.get("/_reload_nginx")
+def show_logs():
+    reload_nginx()
+
+
 ################ DOMAINS API ####################
 
 @app.get("/domain_ping", response_class=PlainTextResponse)
@@ -199,7 +206,7 @@ def ping():
     return "OK"
 
 
-@app.post("/domain_update_nginx_and_ssl")
+@app.get("/domain_update_nginx_and_ssl")
 async def create_nginx(request: Request, domain_id: int):
     pg = request.app.state.pg
 
@@ -213,30 +220,83 @@ async def create_nginx(request: Request, domain_id: int):
     if not row:
         raise HTTPException(status_code=404, detail="Domain not found")
 
+    async with pg.acquire() as conn:
+        await conn.fetchrow("""
+            UPDATE domains
+            SET updated_at = NOW(), ssl_status= 'pending'
+            WHERE id = $1
+        """, domain_id)
+
     domain = row["domain"]
-    path = generate_nginx_conf(domain, domain_id)
+    path = await generate_nginx_conf(domain, domain_id)
+
+    if path:
+        async with pg.acquire() as conn:
+            await conn.fetchrow("""
+                UPDATE domains
+                SET updated_at = NOW(), ssl_status= 'success'
+                WHERE id = $1
+            """, domain_id)
+
     return {"status": "ok", "file": str(path)}
 
+async def request_ssl_letsencrypt(domain: str) -> bool:
+    try:
+        email = f"admin@{domain}"
 
-def generate_nginx_conf(domain: str, domain_id: int) -> Path:
-    template_path = Path("../nginx/nginx.prod.conf")
-    output_dir = Path("../nginx/domains")
+        result = subprocess.run([
+            "certbot", "certonly", "--webroot",
+            "-w", "/var/www/certbot",  # корень для ACME challenge
+            "-d", domain,
+            "--agree-tos",
+            "--email", email,
+            "--non-interactive"
+        ], check=True)
+
+        await asyncio.sleep(20)
+
+        cert_path = Path(f"/etc/letsencrypt/live/{domain}/fullchain.pem")
+        return cert_path.exists()
+
+    except subprocess.CalledProcessError as e:
+        print(f"❌ Certbot failed for {domain}: {e}")
+        return False
+
+
+def reload_nginx():
+    # subprocess.run(["docker", "stop", "tracker_nginx"], check=True)
+    # subprocess.run(["docker", "start", "tracker_nginx"], check=True)
+    subprocess.run([
+        "docker", "exec", "tracker_nginx", "nginx", "-s", "reload"
+    ])
+
+
+async def generate_nginx_conf(domain: str, domain_id: int) -> Path:
+    template_path = Path("/var/www/nginx/nginx.prod.conf")
+    output_dir = Path("/var/www/nginx/domains")
 
     if not template_path.exists():
         raise FileNotFoundError(f"Template file not found: {template_path}")
 
-    # Читаем шаблон
-    content = template_path.read_text()
+    if await request_ssl_letsencrypt(domain):
+        # Читаем шаблон
+        content = template_path.read_text()
 
-    # Заменяем плейсхолдер
-    updated = content.replace("yourdomain.com", domain)
+        # Заменяем плейсхолдер
+        updated = content.replace("server_name _;", f"server_name {domain};")
+        updated = updated.replace("yourdomain.com", domain)
 
-    # Создаём целевой путь
-    output_dir.mkdir(parents=True, exist_ok=True)
-    target_path = output_dir / f"{domain}_{domain_id}.conf"
+        # Создаём целевой путь
+        output_dir.mkdir(parents=True, exist_ok=True)
+        target_path = output_dir / f"{domain_id}_{domain}.conf"
 
-    # Записываем результат
-    target_path.write_text(updated)
-    print(f"✅ Сконфигурирован: {target_path}")
+        if not target_path.exists() or target_path.read_text() != updated:
+            target_path.write_text(updated)
+            reload_nginx()
+            print(f"✅ Updated: {target_path}")
+        else:
+            print(f"ℹ️ No changes: {target_path}")
+    else:
+        raise HTTPException(status_code=500, detail="SSL certificate request failed")
 
     return target_path
